refactor(editor): move log type guards to logs.types
