+++ packages/frontend/editor-ui/src/features/execution/logs/components/LogDetailsPanel.test.ts
@@ -15,9 +15,9 @@ import {
 	mockLoadedNodeType,
 } from '@/__tests__/mocks';
 import { LOG_DETAILS_PANEL_STATE } from '@/features/execution/logs/logs.constants';
-import type { GroupLogEntry, NodeLogEntry } from '../logs.types';
+import { type GroupLogEntry, type NodeLogEntry, isGroupLog } from '../logs.types';
 import { createTestLogEntry } from '../__test__/mocks';
-import { createLogTree, isGroupLog } from '../logs.utils';
+import { createLogTree } from '../logs.utils';
 import { createRunExecutionData, NodeConnectionTypes } from 'n8n-workflow';
 import { HTML_NODE_TYPE } from '@/app/constants';
 import { MESSAGE_AN_AGENT_NODE_TYPE } from '@/app/constants/nodeTypes';
