+++ packages/frontend/editor-ui/src/features/execution/logs/components/LogsPanel.vue
@@ -10,8 +10,7 @@ import { injectNDVStore } from '@/features/ndv/shared/ndv.store';
 import { ndvEventBus } from '@/features/ndv/shared/ndv.eventBus';
 import { useLogsSelection } from '@/features/execution/logs/composables/useLogsSelection';
 import { useLogsTreeExpand } from '@/features/execution/logs/composables/useLogsTreeExpand';
-import { type LogEntry } from '@/features/execution/logs/logs.types';
-import { isNodeLog } from '@/features/execution/logs/logs.utils';
+import { type LogEntry, isNodeLog } from '@/features/execution/logs/logs.types';
 import { useLogsStore } from '@/app/stores/logs.store';
 import { useLogsPanelLayout } from '@/features/execution/logs/composables/useLogsPanelLayout';
 import { type KeyMap } from '@/app/composables/useKeybindings';
