+++ packages/frontend/editor-ui/src/features/ndv/runData/components/ai/RunDataAi.vue
@@ -2,8 +2,8 @@
 import LogsOverviewRows from '@/features/execution/logs/components/LogsOverviewRows.vue';
 import { useLogsExecutionData } from '@/features/execution/logs/composables/useLogsExecutionData';
 import { useLogsTreeExpand } from '@/features/execution/logs/composables/useLogsTreeExpand';
-import type { LogEntry, LogTreeFilter } from '@/features/execution/logs/logs.types';
-import { findLogEntryById, isNodeLog } from '@/features/execution/logs/logs.utils';
+import { type LogEntry, type LogTreeFilter, isNodeLog } from '@/features/execution/logs/logs.types';
+import { findLogEntryById } from '@/features/execution/logs/logs.utils';
 import type { INodeUi } from '@/Interface';
 import { N8nText } from '@n8n/design-system';
 import { useI18n } from '@n8n/i18n';
